requirements.txt
appdirs==1.4.3
Django==1.10.6
django-simple-captcha==0.5.5
django-tinymce==2.6.0
DjangoUeditor==1.8.143
olefile==0.44
packaging==16.8
Pillow==4.1.0
pyparsing==2.2.0
six==1.10.0
